opengp::glfwTrackball inline
opengp::glfwTrackball is defined in a header and was nonetheless not marked inline thus giving it external linkage.  This caused link time errors (multiple definitions of the same symbol) when linking a program or library which included the header containing opengp::glfwTrackball in multiple translation units.

Marked opengp::glfwTrackball inline to rectify this.